REFACTORED: Code that passes Book model to pagination instead of Book Dto and also changes the book id type from int to long in wishlist model class

+++ src/main/java/com/bridgelabz/bookstore/model/Wishlist.java
@@ -10,22 +10,22 @@ public class Wishlist {
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
-    private int userId;
-    private int bookId;
+    private long userId;
+    private long bookId;
     
     public Wishlist() {
     }
  
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
-	public int getBookId() {
+	public long getBookId() {
 		return bookId;
 	}
-	public void setBookId(int bookId) {
+	public void setBookId(long bookId) {
 		this.bookId = bookId;
 	}
